pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.3",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "celery>=5.5.3",
    "fastapi>=0.116.0",
    "loguru>=0.7.3",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
